Consolidate into one method on the ComputeEngineComputerLauncher.
