+++ src/main/java/com/google/jenkins/plugins/computeengine/ComputeEngineLinuxLauncher.java
@@ -24,12 +24,9 @@ import com.google.jenkins.plugins.computeengine.ssh.GoogleKeyPair;
 import com.trilead.ssh2.Connection;
 import com.trilead.ssh2.HTTPProxyData;
 import com.trilead.ssh2.ServerHostKeyVerifier;
-import com.trilead.ssh2.Session;
 import hudson.ProxyConfiguration;
 import hudson.model.TaskListener;
-import hudson.remoting.Channel;
 import java.io.IOException;
-import java.io.PrintStream;
 import java.net.InetSocketAddress;
 import java.net.Proxy;
 import java.util.Optional;
@@ -49,51 +46,6 @@ public class ComputeEngineLinuxLauncher extends ComputeEngineComputerLauncher {
       String cloudName, Operation insertOperation, boolean useInternalAddress) {
     super(cloudName, insertOperation.getName(), insertOperation.getZone());
     this.useInternalAddress = useInternalAddress;
-  }
-
-  @Override
-  protected void launch(ComputeEngineComputer computer, TaskListener listener, Instance inst) {
-    // TODO(#96): Conslidate duplicated launch logic
-    ComputeEngineInstance node = computer.getNode();
-    if (node == null) {
-      logWarning(computer, listener, "Could not get node from computer");
-      return;
-    }
-
-    final Connection conn;
-    Optional<Connection> cleanupConn;
-    PrintStream logger = listener.getLogger();
-    logInfo(computer, listener, "Launching instance: " + node.getNodeName());
-    try {
-      cleanupConn = setupConnection(node, computer, listener);
-      if (!cleanupConn.isPresent()) {
-        return;
-      }
-      conn = cleanupConn.get();
-      String javaExecPath = node.getJavaExecPathOrDefault();
-      if (!checkJavaInstalled(computer, conn, logger, listener, javaExecPath)) {
-        return;
-      }
-      String jenkinsDir = node.getRemoteFS();
-      copyAgentJar(computer, conn, listener, jenkinsDir);
-      String launchString = getJavaLaunchString(javaExecPath, jenkinsDir);
-      logInfo(computer, listener, "Launching Jenkins agent via plugin SSH: " + launchString);
-      final Session sess = conn.openSession();
-      sess.execCommand(launchString);
-      computer.setChannel(
-          sess.getStdout(),
-          sess.getStdin(),
-          logger,
-          new Channel.Listener() {
-            @Override
-            public void onClosed(Channel channel, IOException cause) {
-              sess.close();
-              conn.close();
-            }
-          });
-    } catch (Exception e) {
-      logException(computer, listener, "Error getting exception", e);
-    }
   }
 
   @Override
@@ -262,6 +214,7 @@ public class ComputeEngineLinuxLauncher extends ComputeEngineComputerLauncher {
     }
   }
 
+  @Override
   protected String getPathSeparator() {
     return "/";
   }
